Create random log name
Even though, the log doesn’t contain critical or personal information, it seemed a good idea to create a random file name for it from now on.

The file name is unique for a given site and it is persistent, so it can be shared. It only changes with the `AUTH_KEY` constant given in `wp-config.php`.