Adjusted autoload for PSR-4 compatibility
This is a maintenance adjustment not needed for production. When telling PHPStorm to use a composer.json file to map namespaces according to psr-4, one can only use one. But our project has two, one for the base code, one for premium features. So adding the namespace of pro to the base composer.json is a quick solution.