More performant query for missing sources
The WP Query I used to determine images without sources was not performant on large sites. I replaced it with a custom query that performed very well in comparision and finshed in a fraction of the time.

The result of the query is already cached in a transient, no further caching needed.